Add TODO after “*.la” files removal.
The status of reverse-dependencies has to be checked. If the removal of
the “*.la” files doesn't lead to FTBFSes of those packages, the Depends:
of the -dev package should be adjusted (so as to match the dependencies
expressed in the “*.pc” files). A rebuild of the reverse-dependencies
might then be a good idea.